requirements.txt
python-binance==1.0.22
pandas==2.2.2
numpy==1.26.4
requests==2.32.3
